requirements.txt
fasttext_wheel==0.9.2
Flask==2.3.2
matplotlib==3.7.2
nltk==3.8.1
numpy==1.24.4
pandas==2.0.3
pywebview==4.2.2
scikit_learn==1.3.0
scipy==1.11.1
sentence_transformers==2.2.2
spacy==3.6.0
torch==2.0.1+cu117
Unidecode==1.3.6
youtube_transcript_api==0.6.1
